Add left icon support to product form fields and enhance styling
- Introduced left icon property in form field configuration.
- Implemented left icon handling in the ProductForm component.
- Added TextWithAddon component for rendering fields with icons.
- Updated styles for form manager price input.
