+++ includes/ProductForm/Component.php
@@ -39,6 +39,7 @@ abstract class Component {
         'title'                  => '', // label for the field
         'description'            => '', // description of the field
         'help_content'           => '', // help content for the field
+        'left_icon'              => '', // left icon for the field
         'tooltip'                => '', // tooltip content for the field
         'visibility'             => true, // field visibility, if the field is visible under frontend
         'required'               => false, // by default, all fields are not required
@@ -175,6 +176,34 @@ abstract class Component {
      */
     public function set_help_content( string $help_content ): self {
         $this->data['help_content'] = wp_kses_post( $help_content );
+
+        return $this;
+    }
+
+    /**
+     * Get field left icon
+     *
+     * @since DOKAN_SINCE
+     *
+     * @return string
+     */
+    public function get_left_icon(): string {
+        return $this->data['left_icon'];
+    }
+
+    /**
+     * Set field left icon, validated with
+     *
+     * @since DOKAN_SINCE
+     *
+     * @param string $left_icon
+     *
+     * @see   wp_kses_post()
+     *
+     * @return $this
+     */
+    public function set_left_icon( string $left_icon ): self {
+        $this->data['left_icon'] = wp_kses_post( $left_icon );
 
         return $this;
     }
